+++ owncloudApp/src/main/java/com/owncloud/android/presentation/files/filelist/MainFileListFragment.kt
@@ -31,6 +31,7 @@ import android.content.Intent
 import android.content.res.ColorStateList
 import android.graphics.drawable.Drawable
 import android.net.Uri
+import android.os.Build
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.Menu
@@ -280,8 +281,7 @@ class MainFileListFragment : Fragment(),
 
         showOrHideFab(requireArguments().getParcelable(ARG_FILE_LIST_OPTION)!!, requireArguments().getParcelable(ARG_INITIAL_FOLDER_TO_DISPLAY)!!)
 
-        binding.fabMain.findViewById<AddFloatingActionButton>(com.getbase.floatingactionbutton.R.id.fab_expand_menu_button).contentDescription =
-            getString(R.string.content_description_add_new_content)
+        setFabMainContentDescription()
 
         setTextHintRootToolbar()
     }
@@ -859,6 +859,11 @@ class MainFileListFragment : Fragment(),
                 fabMkdir.isFocusable = isFabExpanded()
                 fabNewfile.isFocusable = isFabExpanded()
                 fabNewshortcut.isFocusable = isFabExpanded()
+                if (fabMain.isExpanded) {
+                    binding.fabMain.findViewById<AddFloatingActionButton>(com.getbase.floatingactionbutton.R.id.fab_expand_menu_button).contentDescription = getString(R.string.content_description_add_new_content_expanded)
+                } else {
+                    setFabMainContentDescription()
+                }
             }
         }
     }
@@ -917,6 +922,11 @@ class MainFileListFragment : Fragment(),
     }
 
     fun isFabExpanded() = binding.fabMain.isExpanded
+
+    fun setFabMainContentDescription() {
+        binding.fabMain.findViewById<AddFloatingActionButton>(com.getbase.floatingactionbutton.R.id.fab_expand_menu_button).contentDescription =
+            getString(R.string.content_description_add_new_content)
+    }
 
     private fun openBottomSheetToUploadFiles() {
         val uploadBottomSheet = layoutInflater.inflate(R.layout.upload_bottom_sheet_fragment, null)
@@ -1411,8 +1421,28 @@ class MainFileListFragment : Fragment(),
                     openInWebProviders = emptyMap()
                 }
             }
+            setRolesAccessibilityToMenuItems()
 
             return true
+        }
+
+        private fun setRolesAccessibilityToMenuItems() {
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+                val roleAccessibilityDescription = getString(R.string.button_role_accessibility)
+                menu?.apply {
+                    findItem(R.id.file_action_select_all)?.contentDescription =  getString(R.string.actionbar_select_all) + roleAccessibilityDescription
+                    findItem(R.id.action_select_inverse)?.contentDescription =  getString(R.string.actionbar_select_inverse) + roleAccessibilityDescription
+                    findItem(R.id.action_open_file_with)?.contentDescription =  getString(R.string.actionbar_open_with) + roleAccessibilityDescription
+                    findItem(R.id.action_rename_file)?.contentDescription =  getString(R.string.common_rename) + roleAccessibilityDescription
+                    findItem(R.id.action_move)?.contentDescription =  getString(R.string.actionbar_move) + roleAccessibilityDescription
+                    findItem(R.id.action_copy)?.contentDescription =  getString(R.string.copy) + roleAccessibilityDescription
+                    findItem(R.id.action_send_file)?.contentDescription = getString(R.string.actionbar_send_file) + roleAccessibilityDescription
+                    findItem(R.id.action_set_available_offline)?.contentDescription =  getString(R.string.set_available_offline) + roleAccessibilityDescription
+                    findItem(R.id.action_unset_available_offline)?.contentDescription =  getString(R.string.unset_available_offline) + roleAccessibilityDescription
+                    findItem(R.id.action_see_details)?.contentDescription =  getString(R.string.actionbar_see_details) + roleAccessibilityDescription
+                    findItem(R.id.action_remove_file)?.contentDescription =  getString(R.string.common_remove) + roleAccessibilityDescription
+                }
+            }
         }
 
         override fun onActionItemClicked(mode: ActionMode?, item: MenuItem?): Boolean {